+++ pyvela/pyvela/convert_components.py
@@ -91,6 +91,8 @@ def pint_components_to_vela(model: TimingModel, toas: TOAs):
 
     if "DMWaveX" in component_names:
         components.append(vl.DMWaveX())
+    elif "PLDMNoiseGP" in component_names:
+        components.append(vl.PowerlawDispersionNoiseGP(int(model.TNDMC.value)))
 
     if "FDJumpDM" in component_names:
         fdjumpdms = list(
@@ -121,6 +123,8 @@ def pint_components_to_vela(model: TimingModel, toas: TOAs):
 
     if "CMWaveX" in component_names:
         components.append(vl.CMWaveX())
+    elif "PLChromNoiseGP" in component_names:
+        components.append(vl.PowerlawChromaticNoiseGP(int(model.TNCHROMC.value)))
 
     if model.BINARY.value is not None:
         assert (model["PB"].quantity is not None) != (model["FB0"].quantity is not None)
@@ -170,6 +174,8 @@ def pint_components_to_vela(model: TimingModel, toas: TOAs):
 
     if "WaveX" in component_names:
         components.append(vl.WaveX())
+    elif "PLRedNoiseGP" in component_names:
+        components.append(vl.PowerlawRedNoiseGP(int(model.TNREDC.value)))
 
     if "Spindown" in component_names:
         components.append(vl.Spindown())