+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Unreleased
 ## Added
 - Check for `ECORR` exclusivity
-- Vectorized option in `get_lnpost_func`
+- Vectorize option in `get_lnpost_func`
+- GP noise models - `PowerlawRedNoiseGP`, `PowerlawDispersionNoiseGP`, `PowerlawChromaticNoiseGP`
+- Examples - `sim3.gp`, `sim4.gp`, `sim6.gp`
 ## Changed
 - Renamed `pint2vela` -> `pyvela`
 - Infer `is_tzr(toa)` from `toa.index`